# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradeVision AI - Enterprise AI Trading Automation Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /strategies:
    get:
      operationId: getStrategies
      tags: [strategies]
      summary: List strategies
      responses:
        "200":
          description: Strategy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    patch:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Strategy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}/deploy:
    post:
      operationId: deployStrategy
      tags: [strategies]
      summary: Deploy strategy to broker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeployInput"
      responses:
        "200":
          description: Deployed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}/backtest-preview:
    get:
      operationId: getStrategyBacktestPreview
      tags: [strategies]
      summary: Get backtest preview for a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Backtest preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestSummary"

  # ─── MARKETPLACE ───────────────────────────────────────────────────────────────
...
  /marketplace/{id}/subscribe:
    post:
      operationId: subscribeToStrategy
      tags: [marketplace]
      summary: Subscribe to a marketplace strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListing"

  # ─── BACKTESTING ───────────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Strategy:
      type: object
      required: [id, name, status, market, symbol, timeframe, description]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, ACTIVE, DEPLOYED]
        market:
          type: string
        symbol:
          type: string
        timeframe:
          type: string
        description:
          type: string
        riskPerTrade:
          type: number
        takeProfit:
          type: number
        stopLoss:
          type: number
        trailingStop:
          type: number
        magicNumber:
          type: ["integer", "null"]
    StrategyInput:
      type: object
      required: [name, market, symbol, timeframe]
      properties:
        name:
          type: string
        market:
          type: string
        symbol:
          type: string
        timeframe:
          type: string
        description:
          type: string
    StrategyUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        riskPerTrade:
          type: number
        takeProfit:
          type: number
        stopLoss:
          type: number
    DeployInput:
      type: object
      required: [accountId, mode]
      properties:
        accountId:
          type: integer
        mode:
          type: string
          enum: [DEMO, LIVE]
    BacktestSummary:
      type: object
      required: [netProfit, totalReturn, profitFactor, winRate, totalTrades, sharpeRatio, maxDrawdown, expectancy]
      properties:
        netProfit:
          type: number
        totalReturn:
          type: number
        profitFactor:
          type: number
        winRate:
          type: number
        totalTrades:
          type: integer
        sharpeRatio:
          type: number
        maxDrawdown:
          type: number
        expectancy:
          type: number

    # ─── MARKETPLACE ─────────────────────────────────────────────────────────
    MarketplaceListing:
      type: object
      required: [id, name, type, description, winRate, profitFactor, maxDrawdown, price, author, rating, subscribers]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: string
        winRate:
          type: number
        profitFactor:
          type: number
        maxDrawdown:
          type: number
        price:
          type: number
        author:
          type: string
        authorRating:
          type: number
        rating:
          type: number
        subscribers:
          type: integer
        isSubscribed:
          type: boolean
        isPremium:
          type: boolean
        isNew:
          type: boolean
        market:
          type: string
        timeframe:
          type: string
...
    Notification:
      type: object
      required: [id, type, title, message, time, isRead]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [risk, bot, trade, system, account, security]
        title:
          type: string
        message:
          type: string
        time:
          type: string
        isRead:
          type: boolean
        badge:
          type: ["string", "null"]
...