# lib/api-zod/src/generated/types/backtestSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TradeVision AI - Enterprise AI Trading Automation Platform
 * OpenAPI spec version: 0.1.0
 */

export interface BacktestSummary {
  netProfit: number;
  totalReturn: number;
  profitFactor: number;
  winRate: number;
  totalTrades: number;
  sharpeRatio: number;
  maxDrawdown: number;
  expectancy: number;
}