# lib/api-zod/src/generated/types/strategyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TradeVision AI - Enterprise AI Trading Automation Platform
 * OpenAPI spec version: 0.1.0
 */

export interface StrategyInput {
  name: string;
  market: string;
  symbol: string;
  timeframe: string;
  description?: string;
}